Rename http.js to networkRequests.js
